Add basic oAuth setup
Add in an oAuth setup to the make page. It will now display a link which
will start an oAuth flow. Once the flow is complete the drop downs will
fill in the relevent options so a user can select the profile they want
to use.

It uses the JavaScript Google oAuth flow so no information about the
user ever gets sent back to the server and it will all be gone on page
refresh. This might want to be changed to cache the user or their
options in the future.

TODO:

- Display a loading indicator while requesting Ajax data
- Handle errors
- Automatically pick the account/property/profile if there is only 1 option
- Add some stylin' so it looks better